backend_db/requirements.txt
numpy>=1.26.4,<2.0.0
 scikit-learn==1.6.1
scipy
librosa
torch
deep-translator
assemblyai
python-dotenv